Edit: Change type of Experience id from long to integer

+++ src/main/java/com/aras/cvcreator/service/ExperienceService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ExperienceService {
     List<Experience> findAll();
 
-    Experience findById(Long id);
+    Experience findById(Integer id);
 
     Experience save(Experience experience);
