+++ src/main/java/com/aras/cvcreator/service/ExperienceServiceImpl.java
@@ -21,7 +21,7 @@ public class ExperienceServiceImpl implements ExperienceService {
     }
 
     @Override
-    public Experience findById(Long id) {
+    public Experience findById(Integer id) {
         return repository.findById(id.intValue()).orElse(null);
     }
